- Receipt cryptographic validation infrastructure - Implemented receipt_crypto.rs (505 LOC) with Ed25519 signature verification, nonce-based replay attack prevention via NonceTracker, and ProviderRegistry for verifying storage/compute provider signatures. Added ValidationError enum covering empty signatures, invalid signatures, unknown providers, and replayed nonces.
- Merkle proof-of-retrievability system - Built storage/src/merkle_proof.rs (382 LOC) implementing binary Merkle tree construction, challenge-proof generation, and verification with DoS protection (max 64-level tree depth). Created storage/tests/proof_security.rs (230 LOC) validating that providers cannot forge proofs without actual data possession.
- Storage provider integration contracts - Implemented client_integration.rs and provider_integration.rs defining StorageContractBuilder, ChallengeResponse, and proof verification workflows. Extended StorageContract struct with storage_root field and verify_proof() method accepting Merkle proofs at specific block heights.
- Governance authorization and access control - Created access_control.rs (428 LOC), authorization.rs (482 LOC), and disbursement_auth.rs (137 LOC) implementing role-based permissions for treasury operations. Added Role, Permission, and AccessPolicy types with multi-sig quorum requirements for disbursements exceeding thresholds.
- Circuit breaker implementation - Built circuit_breaker.rs (562 LOC) with state machine (Closed/Open/HalfOpen), exponential backoff, and telemetry integration. Added circuit_breaker_integration_test.rs validating failure thresholds, automatic recovery, and metrics emission for treasury_circuit_breaker_state.
- Consensus telemetry expansion - Implemented consensus_metrics.rs, consensus_instrumentation.rs, and consensus_integration.rs (561 LOC total) exposing BLOCK_HEIGHT, FORK_DETECTED, MEMPOOL_SIZE, FINALITY_LAG, RECEIPT_VALIDATION_TIME, and STORAGE_PROOF_VALIDATION_TIME metrics.
- Treasury and energy market telemetry - Extended telemetry/treasury.rs (349 LOC) and telemetry/energy.rs (315 LOC) with TREASURY_BALANCE, DISBURSEMENT_COUNT, ENERGY_VERIFICATIONS_TOTAL, and ENERGY_SETTLEMENT_AMOUNT gauges/counters. Added Prometheus recording rules for rate calculations and SLO tracking.
- Foundation serde derive refactoring - Restructured foundation_serde_derive with ast.rs, parser.rs, tokens.rs, error.rs, and generics.rs modules (776 LOC added) improving macro hygiene and error messages. Refactored de.rs and ser.rs (~1200 LOC changed) to use new AST representation.
Receipt stress and security tests - Expanded receipt_stress_tests.rs (+156 LOC) with concurrent validation scenarios. Created receipt_security.rs (420 LOC) and security_integration_tests.rs (58 LOC) testing signature forgery, nonce replay, and unauthorized provider attacks.
- Treasury database migration tooling - Built explorer/src/bin/migrate_treasury_db.rs (62 LOC) for SQLite schema migrations of treasury tables. Updated explorer/src/lib.rs (-32 LOC) removing deprecated field accessors and fixing query bindings.
- Test harness improvements - Added httpd/tests/support.rs (35 LOC) with shared test utilities. Enhanced WebSocket test infrastructure in runtime/tests/ws_shared.rs (+24 LOC) and ws_server.rs for concurrent connection scenarios.
- Documentation and operational guides - Created SECURITY_DEPLOYMENT.md (426 LOC), ENERGY_RPC_ENDPOINTS.md (548 LOC), and TREASURY_RPC_ENDPOINTS.md (469 LOC) documenting API surfaces. Updated operations.md with circuit breaker runbooks and alert correlation matrices.

+++ crates/sys/src/lib.rs
@@ -459,7 +459,7 @@ pub mod kqueue {
             let mut change = RawKevent::zeroed();
             change.ident = ident as usize;
             change.filter = EVFILT_VNODE;
-            change.flags = (EV_ADD | EV_ENABLE | EV_CLEAR) as u16;
+            change.flags = EV_ADD | EV_ENABLE | EV_CLEAR;
             change.fflags = flags;
             change.data = 0;
             change.set_udata_null();
